+++ emachinery/acmsimcv5/c/pid_regulator.h
@@ -49,6 +49,6 @@ extern PID_REG pid1_ic;
 
 void ACMSIMC_PIDTuner();
 
-void commands(double *p_rpm_speed_command, double *p_amp_current_command);
+void commands(REAL *p_rpm_speed_command, REAL *p_amp_current_command);
 
 #endif
